Make webservices secure

+++ src/main/java/be/luxuryoverdosis/framework/business/webservice/rest/GetUserRest.java
@@ -2,7 +2,6 @@ package be.luxuryoverdosis.framework.business.webservice.rest;
 
 import java.util.ArrayList;
 
-import org.apache.commons.lang.StringUtils;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -11,6 +10,7 @@ import org.springframework.web.bind.annotation.RestController;
 import be.luxuryoverdosis.framework.business.service.BaseSpringServiceLocator;
 import be.luxuryoverdosis.framework.business.service.interfaces.RoleService;
 import be.luxuryoverdosis.framework.business.service.interfaces.UserService;
+import be.luxuryoverdosis.framework.business.thread.ThreadManager;
 import be.luxuryoverdosis.framework.data.dto.UserDTO;
 import be.luxuryoverdosis.framework.data.to.Role;
 import be.luxuryoverdosis.framework.data.to.User;
@@ -20,42 +20,39 @@ import net.sf.json.JSONArray;
 @RestController
 @RequestMapping("/user")
 public class GetUserRest {
-	String message;
-	
 	@RequestMapping(value = "/readUserRequest", method = RequestMethod.GET)
 	//public String readUserRequest(@RequestParam(value="name") String name) throws Exception {
 	public String readUserRequest(@RequestHeader(value="name") String name) throws Exception {
+		if(ThreadManager.getUserFromThread() == null) {
+			return BaseSpringServiceLocator.getMessage("security.access.denied");
+		}
+		
 		UserDTO userDTO = null;
 		User user = getUserService().readName(name);
 		if (user == null) {
-			message = BaseSpringServiceLocator.getMessage("exists.not", new Object[]{BaseSpringServiceLocator.getMessage("table.user")});
-		} else {
-			userDTO = getUserService().readDTO(user.getId());
-			if (userDTO == null) {
-				message = BaseSpringServiceLocator.getMessage("exists.not", new Object[]{BaseSpringServiceLocator.getMessage("table.user")});
-			}
+			return  BaseSpringServiceLocator.getMessage("exists.not", new Object[]{BaseSpringServiceLocator.getMessage("table.user")});
+		} 
+		
+		userDTO = getUserService().readDTO(user.getId());
+		if (userDTO == null) {
+			return BaseSpringServiceLocator.getMessage("exists.not", new Object[]{BaseSpringServiceLocator.getMessage("table.user")});
 		}
 		
-		if (message == null)  {
-			return JSONArray.fromObject(userDTO).toString();
-		}
-		
-		return message;
+		return JSONArray.fromObject(userDTO).toString();
 	}
 	
 	@RequestMapping(value = "/readAllUsersRequest", method = RequestMethod.GET)
 	public String readUserRequest() throws Exception {
+		if(ThreadManager.getUserFromThread() == null) {
+			return BaseSpringServiceLocator.getMessage("security.access.denied");
+		}
 		
 		ArrayList<UserDTO> userDTOList = getUserService().listDTO();
 		if (userDTOList == null) {
-			message = BaseSpringServiceLocator.getMessage("exists.not", new Object[]{BaseSpringServiceLocator.getMessage("table.user")});
+			return BaseSpringServiceLocator.getMessage("exists.not", new Object[]{BaseSpringServiceLocator.getMessage("table.user")});
 		}
 		
-		if (message == null)  {
-			return JSONArray.fromObject(userDTOList).toString();
-		}
-		
-		return message;
+		return JSONArray.fromObject(userDTOList).toString();
 	}
 
 	@RequestMapping(value = "/createOrUpdateUserRequest", method = RequestMethod.POST)
@@ -64,8 +61,12 @@ public class GetUserRest {
 			@RequestHeader(value="encryptedPassword") String encryptedPassword,
 			@RequestHeader(value="email") String email,
 			@RequestHeader(value="roleName") String roleName) throws Exception {
+		
+		if(ThreadManager.getUserFromThread() == null) {
+			return BaseSpringServiceLocator.getMessage("security.access.denied");
+		}
+		
 		boolean isNew = false;
-		String message = StringUtils.EMPTY;
 		
 		User user = getUserService().readName(name);
 		if (user == null) {
@@ -81,35 +82,35 @@ public class GetUserRest {
 		
 		Role role = getRoleService().readName(roleName);
 		if (role == null) {
-			message = BaseSpringServiceLocator.getMessage("exists.not", new Object[]{BaseSpringServiceLocator.getMessage("table.role")});
+			return BaseSpringServiceLocator.getMessage("exists.not", new Object[]{BaseSpringServiceLocator.getMessage("table.role")});
 		}
 		user.setRole(role);
 		
 		try {
 			getUserService().createOrUpdate(user);
 			if (isNew) {
-				message = BaseSpringServiceLocator.getMessage("save.success", new Object[]{BaseSpringServiceLocator.getMessage("table.user")});
+				return BaseSpringServiceLocator.getMessage("save.success", new Object[]{BaseSpringServiceLocator.getMessage("table.user")});
 			} else {
-				message = BaseSpringServiceLocator.getMessage("update.success", new Object[]{BaseSpringServiceLocator.getMessage("table.user")});
+				return BaseSpringServiceLocator.getMessage("update.success", new Object[]{BaseSpringServiceLocator.getMessage("table.user")});
 			}
 		} catch (ServiceException e) {
-			message = e.getMessage();
+			return e.getMessage();
 		}
-		
-		return message;
 	}
 	
 	@RequestMapping(value = "/deleteUserRequest", method = RequestMethod.POST)
 	public String deleteUserRequest(@RequestHeader(value="name") String name) {
+		if(ThreadManager.getUserFromThread() == null) {
+			return BaseSpringServiceLocator.getMessage("security.access.denied");
+		}
+		
 		User user = getUserService().readName(name);
 		if(user != null) {
 			getUserService().delete(user.getId());
-			message = BaseSpringServiceLocator.getMessage("delete.success", new Object[]{BaseSpringServiceLocator.getMessage("table.user")});
+			return BaseSpringServiceLocator.getMessage("delete.success", new Object[]{BaseSpringServiceLocator.getMessage("table.user")});
 		} else {
-			message = BaseSpringServiceLocator.getMessage("exists.not", new Object[]{BaseSpringServiceLocator.getMessage("table.user")});
+			return BaseSpringServiceLocator.getMessage("exists.not", new Object[]{BaseSpringServiceLocator.getMessage("table.user")});
 		}
-		
-		return message;
 	}
 	
 	private UserService getUserService() {
